Revert "feat: add commonjs and esm support with vite"
