+++ .npmignore
@@ -11,4 +11,3 @@ node_modules
 prettier*
 jest.config.js
 tsconfig.json
-vite.config.ts
