+++ CHANGELOG.md
@@ -5,12 +5,6 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.0] - 2023-10-22
-
-### Added
-
-* Now we have EcmaScript modules and CommonJS support!
-
 ## [0.0.6] - 2023-10-15
 
 ### Added
