--- index.d.ts
@@ -1,2 +1,0 @@
-/// <reference path="./types/auth.d.ts" />
-/// <reference path="./types/errors.d.ts" />
