+++ jest.config.js
@@ -1,4 +1,4 @@
-export default {
+module.exports = {
   preset: "ts-jest",
   testEnvironment: "node",
   moduleNameMapper: {
