+++ package.json
@@ -1,69 +1,36 @@
 {
   "name": "meli-node-sdk-unnoficial",
-  "version": "0.1.0",
+  "version": "0.0.6",
   "description": "A Mercadolibre NodeJS SDK with the goal of be the most updated SKD available for NodeJS",
-  "type": "module",
-  "main": "./dist/index.js",
-  "module": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "files": [
-    "dist"
-  ],
+  "main": "dist/index.js",
   "scripts": {
-    "build": "tsc && vite build",
-    "preview": "vite preview",
     "format": "prettier --write .",
-    "tsc": "tsc",
+    "build": "tsc",
     "prepare": "npm run build",
     "test": "jest --no-cache --runInBand --detectOpenHandles --passWithNoTests",
     "test:watch": "jest --no-cache --runInBand --detectOpenHandles --passWithNoTests --watch"
   },
+  "types": "dist/index.d.ts",
   "dependencies": {
     "axios": "1.5.1"
   },
   "devDependencies": {
-    "@rollup/plugin-inject": "^5.0.5",
-    "@types/jest": "^29.5.6",
-    "@types/node": "^20.8.7",
-    "@typescript-eslint/eslint-plugin": "6.8.0",
+    "@types/jest": "^29.5.5",
+    "@types/node": "^20.8.6",
+    "@typescript-eslint/eslint-plugin": "6.7.5",
     "axios-mock-adapter": "^1.22.0",
-    "eslint": "8.52.0",
+    "eslint": "8.51.0",
     "eslint-config-prettier": "9.0.0",
     "eslint-config-standard-with-typescript": "39.1.1",
     "eslint-plugin-import": "2.28.1",
-    "eslint-plugin-n": "16.2.0",
+    "eslint-plugin-n": "16.0.0",
     "eslint-plugin-prettier": "5.0.1",
     "eslint-plugin-promise": "6.1.1",
     "husky": "8.0.3",
     "jest": "29.7.0",
     "prettier": "3.0.3",
     "ts-jest": "29.1.1",
-    "typescript": "5.2.2",
-    "vite": "^4.5.0",
-    "vite-plugin-dts": "^3.6.0",
-    "vite-plugin-node-polyfills": "^0.15.0"
-  },
-  "exports": {
-    ".": {
-      "import": "./dist/index.js",
-      "require": "./dist/index.cjs",
-      "types": "./dist/index.d.ts"
-    },
-    "./auth": {
-      "import": "./dist/auth.js",
-      "require": "./dist/auth.cjs",
-      "types": "./dist/auth.d.ts"
-    },
-    "./common": {
-      "import": "./dist/common.js",
-      "require": "./dist/common.cjs",
-      "types": "./dist/common.d.ts"
-    },
-    "./errors": {
-      "import": "./dist/errors.js",
-      "require": "./dist/errors.cjs",
-      "types": "./dist/errors.d.ts"
-    }
+    "typescript": "5.2.2"
   },
   "engines": {
     "node": ">=18.0.0",
