--- src/auth/index.ts
@@ -1,2 +1,0 @@
-export { default as MercadolibreAPIAuth } from './MercadolibreAPIAuth';
-
