--- src/common/client.ts
@@ -1,7 +1,0 @@
-import axios from "axios";
-
-export default function createAxios() {
-  return axios.create({
-    baseURL: `https://api.mercadolibre.com`,
-  });
-}
