--- src/common/enums.ts
@@ -1,11 +1,0 @@
-export enum GrantTypeEnum {
-  CODE = "code",
-  AUTHORIZATION_CODE = "authorization_code",
-  REFRESH_TOKEN = "refresh_token",
-}
-
-export enum ScopeEnum {
-  OFFLINE_ACCESS = "offline_access",
-  READ = "read",
-  WRITE = "write",
-}
