--- src/common/index.ts
@@ -1,5 +1,0 @@
-export { default as createClient } from "./client";
-export * from "./constants";
-export * from "./enums";
-export * from "./interfaces";
-export * from "./types";
