--- src/common/types.ts
@@ -1,9 +1,0 @@
-export type AuthenticationParams =
-  | { redirectUri?: string; state?: string; pkce?: false }
-  | {
-      redirectUri?: string;
-      state?: string;
-      pkce: true;
-      codeChallenge: string;
-      codeChallengeMethod: "S256" | "plain";
-    };
