+++ src/errors.ts
@@ -1,4 +1,4 @@
-export default class MeliError extends Error {
+export class MeliError extends Error {
   public raw: string | null;
   public status: number | null;
   public reason: string;
@@ -8,5 +8,12 @@ export default class MeliError extends Error {
     this.name = "MeliError";
     this.raw = rawError ?? null;
     this.status = status ?? null;
+  }
+}
+
+export class MeliValidationError extends MeliError {
+  constructor(reason: string, rawError?: string) {
+    super(reason, rawError);
+    this.name = "MeliValidationError";
   }
 }
