--- src/errors/MeliValidationError.ts
@@ -1,8 +1,0 @@
-import MeliError from "./MeliError";
-
-export default class MeliValidationError extends MeliError {
-  constructor(reason: string, rawError?: string) {
-    super(reason, rawError);
-    this.name = "MeliValidationError";
-  }
-}
