--- src/errors/index.ts
@@ -1,2 +1,0 @@
-export { default as MeliError } from "./MeliError";
-export { default as MeliValidationError } from "./MeliValidationError";
