--- types/auth.d.ts
@@ -1,2 +1,0 @@
-export * from '../dist/auth';
-
